Resolved #7 with byte packing A single byte stores the 2-bit weight of all three channels. This will nicely fit into either the 2KiB of SRAM or 1KiB of EEPROM However, a new problem has shown up... I can seem to read the step memory in some places but cannot read it in other places. Something is likely wrong with the conversion. Keeping this issue open as I investigate. Also note that all EEPROM code is currently disabled as it didn't seem to work (or, again, maybe the conversion was broken)

+++ src/multiplexer.h
@@ -12,7 +12,7 @@
 #define record_button        B110 // 6
 #define record_button_master B111 // 7
 
-int getTrigerMux(char channels)
+uint16_t getTrigerMux(char channels)
 {
   digitalWrite(mux_select_0, bitRead(channels, 0));
   digitalWrite(mux_select_1, bitRead(channels, 1));
